+++ src/ert/plugins/hook_implementations/forward_model_steps.py
@@ -422,6 +422,39 @@ and then replace the existing file.
         )
 
 
+class MoveDirectory(ForwardModelStepPlugin):
+    def __init__(self) -> None:
+        super().__init__(
+            name="MOVE_DIRECTORY",
+            command=[
+                str(
+                    (
+                        Path(__file__)
+                        / "../../../resources/shell_scripts/move_directory.py"
+                    ).resolve()
+                ),
+                "<FROM>",
+                "<TO>",
+            ],
+        )
+
+    @staticmethod
+    def documentation() -> Optional[ForwardModelStepDocumentation]:
+        return ForwardModelStepDocumentation(
+            category="utility.file_system",
+            description="""
+The :code:`MOVE_DIRECTORY` job will move directory to target directory.
+If directory already exists, this job will move directory to the target directory
+and then replace the existing directory.
+""",
+            examples="""
+            .. code-block:: bash
+
+                FORWARD_MODEL MOVE_DIRECTORY(<FROM>=dir/to/move,<TO>=to/new/path)
+""",
+        )
+
+
 class Symlink(ForwardModelStepPlugin):
     def __init__(self) -> None:
         super().__init__(
@@ -529,6 +562,7 @@ _UpperCaseFMSteps: List[Type[ForwardModelStepPlugin]] = [
     MakeDirectory,
     MakeSymlink,
     MoveFile,
+    MoveDirectory,
     Symlink,
     TemplateRender,
 ]